Crate rsp2_array_types

Re-exports

pub use crate::vee::dot;

pub use crate::mat::inv;

Modules

mat
Small fixed-size matrix types, compatible with V2/V3/V4

vee

Structs

M2
A linear algebra dense matrix with 2 rows and fixed width.

M3
A linear algebra dense matrix with 3 rows and fixed width.

M4
A linear algebra dense matrix with 4 rows and fixed width.

V2
A 2-dimensional vector with operations for linear algebra.

V3
A 3-dimensional vector with operations for linear algebra.

V4
A 4-dimensional vector with operations for linear algebra.

Traits

Envee
Zero-cost transformations from sequences of arrays into sequences of Vn.

Unvee
Zero-cost transformations from sequences of Vn into sequences of arrays.

Type Definitions

Iter

IterMut

M22
A square dense 2x2 matrix.

M33
A square dense 3x3 matrix.

M44
A square dense 4x4 matrix.