Module rsp2_array_types::mat

Small fixed-size matrix types, compatible with V2/V3/V4

This library primarily uses a row-based formalism; matrices are conceptually understood to be containers of row-vectors. This formalism is most useful when most vectors used are row vectors (in which case most matrix-vector multiplication has the matrix on the right)

Traits

Det
Implementation detail of the inherent method {M22,M33}::det.

FromDiag
Implementation detail of the inherent method {M2,M3,M4}::from_diag.

FromFn
Implementation detail of the free function mat::from_fn.

IntoArray
Implementation detail of the inherent method {M2,M3,M4}::into_array.

IntoMatrix
Implementation detail of the free function mat::from_array.

Inv
Implementation detail of the inherent method {M22,M33}::inv.

IsMatrix
Implementation detail of some free functions that defer to external traits.

Transpose
Implementation detail of the inherent method {M2,M3,M4}::t.

Functions

from_array
Construct a matrix from a 2D array (of rows).

inv
Matrix inverse.

Type Definitions

ArrayT

DetT
Output of det. Probably a scalar type.

DiagT
Argument to from_diag. Probably a vector type like V3<X>.

TransposeT
Output of transpose. Probably a matrix with the dimensions flipped.