Trait rsp2_array_types::mat::IsMatrix

pub trait IsMatrix: Sized { }

Implementation detail of some free functions that defer to external traits.

Fuggedaboudit.

Its purpose is to prevent those functions from producing non-matrix types.