Struct rsp2_array_types::M3

pub struct M3<V>(pub [V; 3]);

A linear algebra dense matrix with 3 rows and fixed width.

Implementations

impl<X> M3<V3<X>>

pub fn eye() -> Self where
Self: One,

Construct the identity matrix.

pub fn from_diag(diag: DiagT<Self>) -> Self where
Self: FromDiag,

Construct a diagonal matrix.

pub fn inv(&self) -> Self where
Self: Inv,

Matrix inverse.

It is recommended to write this as M33::inv(m) or etc. rather than m.inv().

pub fn det(&self) -> DetT<Self> where
Self: Det,

Matrix determinant.

impl<V> M3<V>

pub fn t(&self) -> TransposeT<Self> where
Self: Transpose,

Matrix transpose. (does not conjugate)

pub fn into_array(self) -> ArrayT<Self> where
Self: IntoArray,

Cast into a plain [[T; m]; n].

pub fn as_array(&self) -> &ArrayT<Self> where
Self: IntoArray,

Cast into a plain &[[T; m]; n].

pub fn as_array_mut(&mut self) -> &mut ArrayT<Self> where
Self: IntoArray,

Cast into a plain &mut [[T; m]; n].

impl<X> M3<V2<X>>

pub fn zero() -> Self where
Self: Zero,

Construct the zero matrix.

pub fn from_fn<B, F>(f: F) -> Self where
Self: FromFn<F>,
F: FnMut(usize, usize) -> B,

Construct a fixed-size matrix from a function on indices.

pub fn map<B, F>(self, f: F) -> M3<V2<B>> where
F: FnMut(X) -> B,

Map each scalar element of a matrix.

pub fn try_map<E, B, F>(self, f: F) -> Result<M3<V2<B>>, E> where
F: FnMut(X) -> Result<B, E>,

Apply a fallible function to each scalar element, with short-circuiting.

pub fn opt_map<B, F>(self, f: F) -> Option<M3<V2<B>>> where
F: FnMut(X) -> Option<B>,

Apply a fallible function to each scalar element, with short-circuiting.

impl<X> M3<V3<X>>

pub fn zero() -> Self where
Self: Zero,

Construct the zero matrix.

pub fn from_fn<B, F>(f: F) -> Self where
Self: FromFn<F>,
F: FnMut(usize, usize) -> B,

Construct a fixed-size matrix from a function on indices.

pub fn map<B, F>(self, f: F) -> M3<V3<B>> where
F: FnMut(X) -> B,

Map each scalar element of a matrix.

pub fn try_map<E, B, F>(self, f: F) -> Result<M3<V3<B>>, E> where
F: FnMut(X) -> Result<B, E>,

Apply a fallible function to each scalar element, with short-circuiting.

pub fn opt_map<B, F>(self, f: F) -> Option<M3<V3<B>>> where
F: FnMut(X) -> Option<B>,

Apply a fallible function to each scalar element, with short-circuiting.

impl<X> M3<V4<X>>

pub fn zero() -> Self where
Self: Zero,

Construct the zero matrix.

pub fn from_fn<B, F>(f: F) -> Self where
Self: FromFn<F>,
F: FnMut(usize, usize) -> B,

Construct a fixed-size matrix from a function on indices.

pub fn map<B, F>(self, f: F) -> M3<V4<B>> where
F: FnMut(X) -> B,

Map each scalar element of a matrix.

pub fn try_map<E, B, F>(self, f: F) -> Result<M3<V4<B>>, E> where
F: FnMut(X) -> Result<B, E>,

Apply a fallible function to each scalar element, with short-circuiting.

pub fn opt_map<B, F>(self, f: F) -> Option<M3<V4<B>>> where
F: FnMut(X) -> Option<B>,

Apply a fallible function to each scalar element, with short-circuiting.