Struct rsp2_array_types::V4

pub struct V4<X = f64>(pub [X; 4]);

A 4-dimensional vector with operations for linear algebra.

Implementations

impl<X> V4<X>

pub fn zero() -> Self where
Self: Zero,

Get a zero vector.

This is also available as the free function vee::zero; this static method just provides an easy way to supply a type hint.

pub fn from_fn<F>(f: F) -> Self where
F: FnMut(usize) -> X,

Construct a fixed-size vector from a function on indices.

This is also available as the free function vee::from_fn; this static method just provides an easy way to supply a type hint.

pub fn try_from_fn<E, F>(f: F) -> Result<Self, E> where
F: FnMut(usize) -> Result<X, E>,

Construct a fixed-size vector from a function on indices.

pub fn dot(&self, other: &Self) -> ScalarT<Self> where
Self: Dot,

Get the inner product of two vectors.

It is recommended you write this as V3::dot(a, b), rather than a.dot(b).

pub fn sqnorm(&self) -> ScalarT<Self> where
Self: Dot,

Get the vector’s squared magnitude.

pub fn norm(&self) -> ScalarT<Self> where
Self: Dot,
ScalarT<Self>: PrimitiveFloat,

Get the vector’s magnitude.

pub fn unit(&self) -> Self where
X: Field + PrimitiveFloat,

Normalize the vector.

pub fn axis_unit(i: usize) -> Self where
Self: Zero,
X: Semiring + PrimitiveSemiring,

Get a basis vector.

pub fn random_unit() -> Self where
Self: RandomUnit,

Generate a randomly-oriented unit vector whose direction comes from a uniform distribution.

pub fn random_unit_with(rng: impl Rng) -> Self where
Self: RandomUnit,

Generate a randomly-oriented unit vector whose direction comes from a uniform distribution.

pub fn angle_to(&self, other: &Self) -> ScalarT<Self> where
X: Semiring + PrimitiveFloat,

Get the shortest angle (as a value in [0, pi]) between this vector and another.

pub fn mul_diag(&self, other: &Self) -> Self where
X: Semiring + PrimitiveSemiring,

Perform elementwise multiplication, or multiplication of a vector by a diagonal matrix.

pub fn par(&self, r: &Self) -> Self where
X: Field + PrimitiveFloat,

Get the part of the vector that is parallel to r.

pub fn perp(&self, r: &Self) -> Self where
X: Field + PrimitiveFloat,

Get the part of the vector that is perpendicular to r.

Be aware that chained calls to perp can have spectacularly bad numerical stability issues; you cannot trust that c.perp(a).perp(b) is even remotely orthogonal to a unless b is orthogonal to a. (for 3d vectors, try c.par(a.cross(b)) instead.)

pub fn map<B, F>(self, f: F) -> V4<B> where
F: FnMut(X) -> B,

Apply a function to each element.

pub fn try_map<E, B, F>(self, f: F) -> Result<V4<B>, E> where
F: FnMut(X) -> Result<B, E>,

Apply a fallible function to each element, with short-circuiting.

pub fn opt_map<B, F>(self, f: F) -> Option<V4<B>> where
F: FnMut(X) -> Option<B>,

Apply a fallible function to each element, with short-circuiting.