Trait rsp2_array_types::Envee

pub unsafe trait Envee {
    type En: ?Sized;
    fn envee(self) -> Self::En
    where
        Self: Sized,
        Self::En: Sized,
    { ... }

    fn envee_ref(&self) -> &Self::En { ... }

    fn envee_mut(&mut self) -> &mut Self::En { ... }
}

Zero-cost transformations from sequences of arrays into sequences of Vn.

Safety

The default impls effectively perform transmute, and some of the generic impls assume that it is safe to perform pointer casts between Self and Self::En. (this may be done even on pointers to pointers, or smart pointers and etc.)

Associated Types

type En: ?Sized

Loading content...
Provided methods

fn envee(self) -> Self::En where
Self: Sized,
Self::En: Sized,

Casts a sequence of arrays into V2/V3/V4s.

fn envee_ref(&self) -> &Self::En

Borrow a sequence of arrays as V2/V3/V4s.

This method exists for the convenience of autoref. (Contrast with (&self).envee())

fn envee_mut(&mut self) -> &mut Self::En

Mutably borrow a sequence of arrays as V2/V3/V4s.

This method exists for the convenience of autoref. (Contrast with (&mut self).envee())

Loading content...