Trait rsp2_array_types::Unvee

pub unsafe trait Unvee {
    type Un: ?Sized;
    fn unvee(self) -> Self::Un
    where
        Self: Sized,
        Self::Un: Sized,
    { ... }

    fn unvee_ref(&self) -> &Self::Un { ... }

    fn unvee_mut(&mut self) -> &mut Self::Un { ... }
}

Zero-cost transformations from sequences of Vn into sequences of arrays.

Safety

The default impls effectively perform transmute, and some of the generic impls assume that it is safe to perform pointer casts between Self and Self::Un. (this may be done even on pointers to pointers, or smart pointers and etc.)

Associated Types

type Un: ?Sized

Loading content...
Provided methods

fn unvee(self) -> Self::Un where
Self: Sized,
Self::Un: Sized,

Casts a sequence of V2/V3/V4s into arrays.

fn unvee_ref(&self) -> &Self::Un

Borrow a sequence of V2/V3/V4s as arrays.

This method exists for the convenience of autoref. (Contrast with (&self).unvee())

fn unvee_mut(&mut self) -> &mut Self::Un

Mutably borrow a sequence of V2/V3/V4s as arrays.

This method exists for the convenience of autoref. (Contrast with (&mut self).unvee())

Loading content...